[Update] 테스트 코드 에러로 인한 Service 수정 및 CheckException 삭제

+++ src/main/java/com/likelion/finalproject/service/CommentService.java
@@ -4,9 +4,12 @@ import com.likelion.finalproject.domain.dto.alarm.AlarmDto;
 import com.likelion.finalproject.domain.dto.alarm.AlarmType;
 import com.likelion.finalproject.domain.dto.comment.CommentDto;
 import com.likelion.finalproject.domain.dto.comment.CommentRequest;
+import com.likelion.finalproject.domain.dto.user.UserRole;
 import com.likelion.finalproject.domain.entity.Comment;
 import com.likelion.finalproject.domain.entity.Post;
 import com.likelion.finalproject.domain.entity.User;
+import com.likelion.finalproject.exception.ErrorCode;
+import com.likelion.finalproject.exception.UserException;
 import com.likelion.finalproject.repository.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,7 +22,8 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class CommentService {
 
-    private final CheckException checkException;
+    private final UserRepository userRepository;
+    private final PostRepository postRepository;
     private final CommentRepository commentRepository;
     private final AlarmRepository alarmRepository;
 
@@ -28,7 +32,8 @@ public class CommentService {
      */
     public Page<CommentDto> printComment(Long postId, Pageable pageable) {
         // 포스트가 DB에 존재하지 않을 경우
-        checkException.checkPost(postId);
+        postRepository.findById(postId)
+                .orElseThrow(() -> new UserException(ErrorCode.POST_NOT_FOUND, ErrorCode.POST_NOT_FOUND.getMessage()));
 
         Page<Comment> comments = commentRepository.findByPostId(postId, pageable);
         Page<CommentDto> commentDtos = comments.map(comment -> CommentDto.builder()
@@ -45,13 +50,16 @@ public class CommentService {
 
     public CommentDto createComment(CommentRequest commentRequest, Long postId, String userName) {
         // 작성자(유저)가 DB에 존재하지 않을 경우
-        User user = checkException.checkUser(userName);
+        User user = userRepository.findByUserName(userName)
+                .orElseThrow(() -> new UserException(ErrorCode.USERNAME_NOT_FOUND, ErrorCode.USERNAME_NOT_FOUND.getMessage()));
 
         // 포스트가 DB에 존재하지 않을 경우
-        Post post = checkException.checkPost(postId);
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new UserException(ErrorCode.POST_NOT_FOUND, ErrorCode.POST_NOT_FOUND.getMessage()));
 
         // 포스트 작성자 DB에 존재하지 않을 경우
-        User postUser = checkException.checkPostUser(post.getUserName());
+        User postUser = userRepository.findByUserName(post.getUserName())
+                .orElseThrow(() -> new UserException(ErrorCode.DATABASE_ERROR, "DB에 포스트 작성자가 존재하지 않습니다."));
 
         Comment savedComment = commentRepository.save(CommentDto.toEntity(commentRequest.getComment(), user, post));
 
@@ -72,16 +80,26 @@ public class CommentService {
 
     public CommentDto updateComment(CommentRequest commentRequest, Long postId, Long id, String userName) {
         // 작성자(유저)가 DB에 존재하지 않을 경우
-        User user = checkException.checkUser(userName);
+        User user = userRepository.findByUserName(userName)
+                .orElseThrow(() -> new UserException(ErrorCode.USERNAME_NOT_FOUND, ErrorCode.USERNAME_NOT_FOUND.getMessage()));
 
         // 포스트가 DB에 존재하지 않을 경우
-        Post post = checkException.checkPost(postId);
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new UserException(ErrorCode.POST_NOT_FOUND, ErrorCode.POST_NOT_FOUND.getMessage()));
 
         // 포스트 작성자 DB에 존재하지 않을 경우
-        User postUser = checkException.checkPostUser(post.getUserName());
+        User postUser = userRepository.findByUserName(post.getUserName())
+                .orElseThrow(() -> new UserException(ErrorCode.DATABASE_ERROR, "DB에 포스트 작성자가 존재하지 않습니다."));
 
         // 댓글 수정 가능 여부 체크
-        Comment comment = checkException.checkComment(id, post.getId(), userName, user);
+        // 댓글이 DB에 존재하지 않을 경우
+        Comment comment = commentRepository.findByIdAndPostId(id, post.getId())
+                .orElseThrow(() -> new UserException(ErrorCode.COMMENT_NOT_FOUND, ErrorCode.COMMENT_NOT_FOUND.getMessage()));
+
+        // 작성자 != 유저, 하지만 유저의 ROLE이 ADMIN이면 수정이나 삭제가 가능하도록
+        if(!comment.getUser().getUserName().equals(userName) && user.getRole().equals(UserRole.USER)) {
+            throw new UserException(ErrorCode.INVALID_PERMISSION, ErrorCode.INVALID_PERMISSION.getMessage());
+        }
 
         comment.setComment(commentRequest.getComment());
 
@@ -98,16 +116,26 @@ public class CommentService {
 
     public Long deleteComment(Long postId, Long id, String userName) {
         // 작성자(유저)가 DB에 존재하지 않을 경우
-        User user = checkException.checkUser(userName);
+        User user = userRepository.findByUserName(userName)
+                .orElseThrow(() -> new UserException(ErrorCode.USERNAME_NOT_FOUND, ErrorCode.USERNAME_NOT_FOUND.getMessage()));
 
         // 포스트가 DB에 존재하지 않을 경우
-        Post post = checkException.checkPost(postId);
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new UserException(ErrorCode.POST_NOT_FOUND, ErrorCode.POST_NOT_FOUND.getMessage()));
 
         // 포스트 작성자 DB에 존재하지 않을 경우
-        User postUser = checkException.checkPostUser(post.getUserName());
+        User postUser = userRepository.findByUserName(post.getUserName())
+                .orElseThrow(() -> new UserException(ErrorCode.DATABASE_ERROR, "DB에 포스트 작성자가 존재하지 않습니다."));
 
         // 댓글 수정 가능 여부 체크
-        Comment comment = checkException.checkComment(id, post.getId(), userName, user);
+        // 댓글이 DB에 존재하지 않을 경우
+        Comment comment = commentRepository.findByIdAndPostId(id, post.getId())
+                .orElseThrow(() -> new UserException(ErrorCode.COMMENT_NOT_FOUND, ErrorCode.COMMENT_NOT_FOUND.getMessage()));
+
+        // 작성자 != 유저, 하지만 유저의 ROLE이 ADMIN이면 수정이나 삭제가 가능하도록
+        if(!comment.getUser().getUserName().equals(userName) && user.getRole().equals(UserRole.USER)) {
+            throw new UserException(ErrorCode.INVALID_PERMISSION, ErrorCode.INVALID_PERMISSION.getMessage());
+        }
 
         commentRepository.delete(comment);
